+++ src/components/OverflowModals/NeedToBeOwnerModal.tsx
@@ -8,7 +8,7 @@ import retinaSikkerhetBom from '../../assets/images/sikkerhet_bom@2x.png'
 import sikkerhetBom from '../../assets/images/sikkerhet_bom.png'
 import { useSettings } from '../../settings/SettingsProvider'
 import { CloseButton } from '../CloseButton/CloseButton'
-import './Modals.scss'
+import './OverflowModals.scss'
 
 interface NeedToBeOwnerModalProps {
     open: boolean
