+++ src/containers/Admin/ShareTab/ShareTab.tsx
@@ -16,8 +16,8 @@ import {
 import { BoardOwnersData, Invite } from '../../../types'
 import { LoginModal } from '../../../components/LoginModal/LoginModal'
 import { LoginCase } from '../../../components/LoginModal/login-modal-types'
-import { RemoveSelfFromTavleModal } from '../../../components/Modals/RemoveSelfFromTavleModal'
-import { NeedToBeOwnerModal } from '../../../components/Modals/NeedToBeOwnerModal'
+import { RemoveSelfFromTavleModal } from '../../../components/OverflowModals/RemoveSelfFromTavleModal'
+import { NeedToBeOwnerModal } from '../../../components/OverflowModals/NeedToBeOwnerModal'
 import { EditableBoardTitle } from './components/EditableBoardTitle'
 import { BoardLink } from './components/BoardLink'
 import { AddNewOwnersInput } from './components/AddNewOwnersInput'
